+++ drink/src/lib.rs
@@ -11,8 +11,6 @@ pub use sandbox::*;
 #[cfg(feature = "session")]
 pub mod session;
 
-use std::sync::{Arc, Mutex};
-
 #[cfg(feature = "session")]
 pub use drink_test_macro::{contract_bundle_provider, test};
 pub use errors::Error;
@@ -24,15 +22,10 @@ use frame_system::EventRecord;
 pub use mock::{mock_message, ContractMock, MessageMock, MockedCallResult, Selector};
 use pallet_contracts::{debug::ExecResult, ExecReturnValue};
 use pallet_contracts_uapi::ReturnFlags;
-use parity_scale_codec::{Decode, Encode};
 /// Export pallets that are used in the minimal runtime.
 pub use {frame_support, frame_system, pallet_balances, pallet_contracts, pallet_timestamp};
 
 pub use crate::runtime::minimal::{self, MinimalRuntime};
-use crate::{
-    errors::MessageResult, mock::MockRegistry,
-    runtime::pallet_contracts_debugging::InterceptingExtT,
-};
 
 /// Alias for `frame-system`'s `RuntimeCall` type.
 pub type RuntimeCall<R> = <R as frame_system::Config>::RuntimeCall;
@@ -49,63 +42,3 @@ pub type EventRecordOf<T> =
 
 /// Default gas limit.
 pub const DEFAULT_GAS_LIMIT: Weight = Weight::from_parts(100_000_000_000, 3 * 1024 * 1024);
-
-/// Runtime extension enabling contract call interception.
-struct MockingExtension<AccountId: Ord> {
-    /// Mock registry, shared with the sandbox.
-    ///
-    /// Potentially the runtime is executed in parallel and thus we need to wrap the registry in
-    /// `Arc<Mutex>` instead of `Rc<RefCell>`.
-    mock_registry: Arc<Mutex<MockRegistry<AccountId>>>,
-}
-
-impl<AccountId: Ord + Decode> InterceptingExtT for MockingExtension<AccountId> {
-    fn intercept_call(
-        &self,
-        contract_address: Vec<u8>,
-        _is_call: bool,
-        input_data: Vec<u8>,
-    ) -> Vec<u8> {
-        let contract_address = Decode::decode(&mut &contract_address[..])
-            .expect("Contract address should be decodable");
-
-        match self
-            .mock_registry
-            .lock()
-            .expect("Should be able to acquire registry")
-            .get(&contract_address)
-        {
-            // There is no mock registered for this address, so we return `None` to indicate that
-            // the call should be executed normally.
-            None => None::<()>.encode(),
-            // We intercept the call and return the result of the mock.
-            Some(mock) => {
-                let (selector, call_data) = input_data.split_at(4);
-                let selector: Selector = selector
-                    .try_into()
-                    .expect("Input data should contain at least selector bytes");
-
-                let result = mock
-                    .call(selector, call_data.to_vec())
-                    .expect("TODO: let the user define the fallback mechanism");
-
-                // Although we don't know the exact type, thanks to the SCALE encoding we know
-                // that `()` will always succeed (we only care about the `Ok`/`Err` distinction).
-                let decoded_result: MessageResult<()> =
-                    Decode::decode(&mut &result[..]).expect("Mock result should be decodable");
-
-                let flags = match decoded_result {
-                    Ok(_) => ReturnFlags::empty(),
-                    Err(_) => ReturnFlags::REVERT,
-                };
-
-                let result: ExecResult = Ok(ExecReturnValue {
-                    flags,
-                    data: result,
-                });
-
-                Some(result).encode()
-            }
-        }
-    }
-}
